Testing for the fact that media elements are treated as replaced elements https://bugs.webkit.org/show_bug.cgi?id=229857
Patch by Frédéric Wang <fwang@igalia.com> on 2021-09-08
Reviewed by Darin Adler.

This is a follow-up of r281979. backwards-text-iterator-basic.html is extended to check that
SimplifiedBackwardsTextIterator emits a comma for a media element. The a11y test from r274810
which verified audio/video elements emit an object replacement character is restored, as it
seems it was removed by mistake and so this case was no longer checked explicitly.
text-marker-previous-next.html is also extended to cover the case of media element.

* accessibility/text-marker/media-emits-object-replacement-expected.txt: Added.
* accessibility/text-marker/media-emits-object-replacement.html: Added.
* accessibility/text-marker/text-marker-previous-next-expected.txt:
* accessibility/text-marker/text-marker-previous-next.html:
* editing/text-iterator/backwards-text-iterator-basic-expected.txt:
* editing/text-iterator/backwards-text-iterator-basic.html:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@282134 268f45cc-cd09-0410-ab3c-d52691b4dbfc