[css-position-sticky] Sticky constraints are calculated incorrectly when scrolling container has padding and borders https://bugs.webkit.org/show_bug.cgi?id=229997
Reviewed by Simon Fraser.

LayoutTests/imported/w3c:

* web-platform-tests/css/css-position/sticky/position-sticky-large-top-2.tentative-expected.html: This test uses
an identity transform to make a div a containing block for absolutely positioned items. This exposes a bug in
WebKit, but position: relative also does this without exposing the bug. The bug for this issue is
https://bugs.webkit.org/show_bug.cgi?id=229999.
* web-platform-tests/css/css-position/sticky/position-sticky-large-top.tentative-expected.html: Ditto.

Source/WebCore:

No new tests. This is covered by two existing WPT tests:
    web-platform-tests/css/css-position/sticky/position-sticky-large-top-2.tentative-expected.html
    web-platform-tests/css/css-position/sticky/position-sticky-large-top.tentative-expected.html

* rendering/RenderBoxModelObject.cpp:
(WebCore::RenderBoxModelObject::computeStickyPositionConstraints const): When the containing block and the scroll container
are the same, fix the calculation of the container content rect. layoutOverflowRect() already does not include the border,
so stop adjusting for that. In addition, instead of simply adjusting the rect location by the padding, actually contract
it so that the bottom and right edges of the rectangle are correct.

LayoutTests:

* TestExpectations: Unskip two newly passing tests.
* platform/ios-wk2/scrollingcoordinator/scrolling-tree/coordinated-frame-expected.txt: Rebaseline.
* platform/ios-wk2/scrollingcoordinator/scrolling-tree/coordinated-frame-gain-scrolling-ancestor-expected.txt: Ditto.
* platform/ios-wk2/scrollingcoordinator/scrolling-tree/coordinated-frame-in-fixed-expected.txt: Ditto.
* platform/ios-wk2/scrollingcoordinator/scrolling-tree/coordinated-frame-lose-scrolling-ancestor-expected.txt: Ditto.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@282138 268f45cc-cd09-0410-ab3c-d52691b4dbfc