run-webkit-archive crashes with dyld error https://bugs.webkit.org/show_bug.cgi?id=228060
Reviewed by Alex Christensen.

Source/JavaScriptCore:

Add a compatibility symbol for the old `WTF::Optional` version of `BackendDispatcher::reportProtocolError` in
versions of Safari shipping with macOS Big Sur and earlier. This has the side effect of not surfacing protocol
errors and instead silently discarding them when crossing this API boundary. Newer versions of macOS ship using
the newer symbol and are therefor not affected and will continue to surface protocol errors to the frontend.

* runtime/SymbolStubsForSafariCompatibility.mm:
(Inspector::BackendDispatcher::reportProtocolError):

Source/WTF:

Bump the maximum macOS version that `HAVE_SAFARI_FOR_WEBKIT_DEVELOPMENT_REQUIRING_EXTRA_SYMBOLS` will be enabled
to provide compatibility with changes made in macOS 12.

* wtf/PlatformHave.h:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@282221 268f45cc-cd09-0410-ab3c-d52691b4dbfc