[JSC] Intl.Locale weekendInfo should list all weekend days instead of range https://bugs.webkit.org/show_bug.cgi?id=230108
Reviewed by Ross Kirsling.

JSTests:

* stress/intl-locale-info.js:
(throw.new.Error):
(let.enGB.new.Intl.Locale.shouldBe): Deleted.
* test262/config.yaml:

Source/JavaScriptCore:

We cannot assume that weekend is contiguous. For example, Burnei's weekend is Friday and Sunday.
This is raised in [1], and our conclusion in Sep-9 meeting is that we should have an array which
includes all the weekend days. The change is merged in [2], and this patch changes our implementation
accordingly.

[1]: https://github.com/tc39/proposal-intl-locale-info/issues/25
[2]: https://github.com/tc39/proposal-intl-locale-info/commit/afb1e269dd698476a2514129235cdad88af60e6f

* runtime/IntlLocale.cpp:
(JSC::IntlLocale::weekInfo):
* runtime/IntlObjectInlines.h:
(JSC::createArrayFromIntVector):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@282257 268f45cc-cd09-0410-ab3c-d52691b4dbfc