[css-flexbox] Add support for self-start & self-end css-align-3 positional alignment properties https://bugs.webkit.org/show_bug.cgi?id=229996
Reviewed by Manuel Rego Casasnovas.

Source/WebCore:

Added support for SelfStart and SelfEnd positional alignment properties from
https://drafts.csswg.org/css-align-3/#positional-values. These two properties
align the alignment subject (flex item) to be flush with the edge of the alignment container (flex line)
corresponding to the alignment subject's start side in the appropriate axis. This allows authors to align
the flex items based on their writing modes instead of the writing mode of the container.

* rendering/RenderFlexibleBox.cpp:
(WebCore::alignmentOffset):
(WebCore::RenderFlexibleBox::alignmentForChild const):

LayoutTests:

* TestExpectations: Removed a couple of tests that are now passing.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@282267 268f45cc-cd09-0410-ab3c-d52691b4dbfc