[ JSC Catalina Release ] wasm.yaml/wasm/references/multitable.js.wasm-collect-continuously (javascriptcore-tests) is a flaky timeout https://bugs.webkit.org/show_bug.cgi?id=229896
The combination of collectContinuously and verifyGC really hammer the poor hardware
during the final two "table insanity" stress tests. I have confirmed that the test
is indeed making forward progress, even after running for 45 minutes, by using logGC.
For this reason, it seems like it is OK to skip verifyGC for this test.

We could have added a vm option isVerifyingGC and disabled only the table insanity test
in this case like we do for low memory, but I think this is good enough.

Reviewed by Mark Lam.

* wasm/references/multitable.js:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@282300 268f45cc-cd09-0410-ab3c-d52691b4dbfc