css/css-transforms/translate-getComputedStyle.html fails https://bugs.webkit.org/show_bug.cgi?id=230178
Reviewed by Antti Koivisto.

LayoutTests/imported/w3c:

The interpolation test appears to regress, but do so because we now correctly output calc()
in computed style (we just serialize the calc incorrectly).

* web-platform-tests/css/css-transforms/animation/translate-composition-expected.txt:
* web-platform-tests/css/css-transforms/animation/translate-interpolation-expected.txt:
* web-platform-tests/css/css-transforms/transforms-support-calc-expected.txt:
* web-platform-tests/css/css-transforms/translate-getComputedStyle-expected.txt:
* web-platform-tests/web-animations/animation-model/animation-types/interpolation-per-property-002-expected.txt:

Source/WebCore:

https://drafts.csswg.org/css-transforms-2/#propdef-translate says that the computed value of
the translate property is "the keyword none or a pair of computed <length-percentage> values
and an absolute length" and has a note saying "Note: The resolved value of the translate
property is the computed value, and thus getComputedStyle() includes percentage values in
its results." so implement that.

Tested by imported/w3c/web-platform-tests/css/css-transforms/translate-getComputedStyle.html

* css/CSSComputedStyleDeclaration.cpp:
(WebCore::computedTranslate):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@282315 268f45cc-cd09-0410-ab3c-d52691b4dbfc