Remove the "core.webKitBranchBuild" prompt from setup-git-clone https://bugs.webkit.org/show_bug.cgi?id=230323
Reviewed by Jonathan Bedard.

* Scripts/webkitpy/tool/commands/setupgitclone.py:
(SetupGitClone.execute):
"core.webKitBranchBuild" is an advanced feature and also doesn't work
at all for Apple folks (leading to a myriad of bizarro build problems).
Folks who want it can use the documentation (UsingGitWithWebKit on the Wiki),
but in general, until it's well supported, it's best if we don't encourage it.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@282482 268f45cc-cd09-0410-ab3c-d52691b4dbfc