Stub out the footprint() API when libpas is in use https://bugs.webkit.org/show_bug.cgi?id=230362
Reviewed by Yusuke Suzuki.

The Gigacage's footprint API is called from some test-only code in JSC, evidently to no
meaningful effect. It's possible for libpas to support such an API but it currently
doesn't. So, stub it out. To my knowledge we don't actually use this API for anything
other than tests, and even tests seem to use it only in the sense that they expect to be
able to call it without crashing.

* bmalloc/Gigacage.cpp:
(Gigacage::footprint):



git-svn-id: http://svn.webkit.org/repository/webkit/trunk@282561 268f45cc-cd09-0410-ab3c-d52691b4dbfc