[RenderTreeUpdater] NULL ptr deref in updateRenderTree https://bugs.webkit.org/show_bug.cgi?id=230581
Patch by Brandon Stewart <brandonstewart@apple.com> on 2021-09-24
Reviewed by Antti Koivisto.

Source/WebCore:

Text element changes are buffered. This can lead to undesirable behavior
when switching a node to a document that is not rendered, and then proceeding
with a rendering update.

If we cannot find a renderer in a node or its ancestors then just give up
instead of returning a document.

Test: fast/dom/Document/clearPendingRenderTreeUpdater.html

* rendering/updating/RenderTreeUpdater.cpp:
(WebCore::findRenderingRoot):
(WebCore::RenderTreeUpdater::commit):
(WebCore::RenderTreeUpdater::createRenderer):
(WebCore::RenderTreeUpdater::textRendererIsNeeded):

LayoutTests:

New regression test to handle case where we trigger a text update,
and then switch the node to a new unrendered document.

* fast/dom/Document/clearPendingRenderTreeUpdater-expected.txt: Added.
* fast/dom/Document/clearPendingRenderTreeUpdater.html: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@283030 268f45cc-cd09-0410-ab3c-d52691b4dbfc