Terminate PCM daemon before and after unit test that uses it https://bugs.webkit.org/show_bug.cgi?id=230191
In r282368 I added a unit test that creates a daemon, uses it, then terminates it.
On some bots it has been timing out.  Once it starts timing out, it probably won't stop because
it won't get to the part of the test that terminates the problematic daemon.

* TestWebKitAPI/Tests/WebKitCocoa/EventAttribution.mm:
(TestWebKitAPI::TEST):



git-svn-id: http://svn.webkit.org/repository/webkit/trunk@283298 268f45cc-cd09-0410-ab3c-d52691b4dbfc