[GPU Process] support rendering Apple Pay buttons https://bugs.webkit.org/show_bug.cgi?id=230648 <rdar://problem/72061985>
Reviewed by Tim Horton.

* rendering/RenderThemeCocoa.mm:
(WebCore::RenderThemeCocoa::paintApplePayButton):
Instead of drawing directly into the `GraphicsContext::platformContext` (which will not be
valid in the WebProcess when DOM rendering happens in the GPUProcess), first draw into a
temporary `ImageBuffer` and then consume it to draw into the actual/used `GraphicsContext`.
While it is possible to create a dedicated display list item for this, we don't want to do
that because `PKDrawApplePayButtonWithCornerRadius` involves dealing with PDFs, which are
not as secure as we'd like for use in the GPUProcess.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@283331 268f45cc-cd09-0410-ab3c-d52691b4dbfc