Authorization header lost on 30x redirects https://bugs.webkit.org/show_bug.cgi?id=230935 <rdar://problem/83689955>
Reviewed by Darin Adler.

Source/WebCore:

CFNetwork drops the Authorization request header in cases of same-origin redirects, which is not as per
the fetch specification [1] and doesn't match the behavior of other browsers.

To address the issue, WebKit adds the Authorization request back in case of a same-origin redirect.

Test: http/tests/fetch/fetch-redirect-same-origin-authorization.html

* platform/network/cf/ResourceHandleCFNet.cpp:
(WebCore::ResourceHandle::willSendRequest):
* platform/network/mac/ResourceHandleMac.mm:
(WebCore::ResourceHandle::willSendRequest):

Source/WebKit:

CFNetwork drops the Authorization request header in cases of same-origin redirects, which is not as per
the fetch specification [1] and doesn't match the behavior of other browsers.

To address the issue, WebKit adds the Authorization request back in case of a same-origin redirect.

[1] https://fetch.spec.whatwg.org/#concept-http-redirect-fetch

* NetworkProcess/cocoa/NetworkDataTaskCocoa.mm:
(WebKit::NetworkDataTaskCocoa::willPerformHTTPRedirection):

LayoutTests:

* http/tests/fetch/fetch-redirect-same-origin-authorization-expected.txt: Added.
* http/tests/fetch/fetch-redirect-same-origin-authorization.html: Added.
* http/tests/fetch/resources/dump-authorization-header.py: Added.
Add layout test coverage.

* http/tests/xmlhttprequest/redirections-and-user-headers.html:
Update existing test to reflect behavior change.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@283565 268f45cc-cd09-0410-ab3c-d52691b4dbfc