TestWebKitAPI.PrivateClickMeasurement.DaemonBasicFunctionality is timing out https://bugs.webkit.org/show_bug.cgi?id=231257
* TestWebKitAPI/Tests/WebKitCocoa/EventAttribution.mm:
Originally reviewed as part of https://bugs.webkit.org/show_bug.cgi?id=231248
this should fix the error that happens when a previous test run has registered
the test AdAttributionDaemon but then failed before it was started, which leaves
launchd in a state of having already loaded a plist for the mach service named
"org.webkit.pcmtestdaemon.service" so it refuses to load another.  The solution is
just to try unloading any previously loaded plist before loading another to clear
state in launchd.



git-svn-id: http://svn.webkit.org/repository/webkit/trunk@283574 268f45cc-cd09-0410-ab3c-d52691b4dbfc