PCM: Allow measurement of links in nested, cross-site iframes https://bugs.webkit.org/show_bug.cgi?id=229204 <rdar://problem/82310386>
Reviewed by Alex Christensen.

Developers have said it would be great to allow measurement of links served in
nested, cross-site iframes on the click source. The click still count as if it
happened on the first party click source site, and any attribution report still
goes to that first party. This change enables the *serving* of measurable links
in nested, cross-site iframes.

The standards issue where this was discussed:
https://github.com/privacycg/private-click-measurement/issues/7

Source/WebCore:

Test: http/tests/privateClickMeasurement/store-private-click-measurement-nested.html

* html/HTMLAnchorElement.cpp:
(WebCore::HTMLAnchorElement::parsePrivateClickMeasurement const):
    Removed the block for nested, cross-site iframes.

LayoutTests:

* http/tests/privateClickMeasurement/resources/nestedTargetLink.html: Added.
* http/tests/privateClickMeasurement/store-private-click-measurement-nested-expected.txt: Added.
* http/tests/privateClickMeasurement/store-private-click-measurement-nested.html: Added.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@283593 268f45cc-cd09-0410-ab3c-d52691b4dbfc